Fix pixel flickering issue on AMD graphics cards
Fix validation issue related to a push constant getting optimized away
Remove usage of uint16_t in shaders since AMD sucks at supporting it
Replace almost all usage of ByteAddressBuffer in shaders with StructuredBuffer
Slightly improve naming of descriptors in shaders to be more standardized

+++ shaders/cModelCulling.cs.hlsl
@@ -143,9 +143,9 @@ uint64_t CalculateSortKey(DrawCall drawCall, DrawCallData drawCallData, Instance
         };
     */
     
-    uint64_t renderPriority = drawCallData.renderPriority;
-    uint64_t invDistanceFromCameraUint = UINT_MAX - (uint)(distanceFromCamera / distanceAccuracy);
-    uint64_t localInstanceID = drawCall.firstInstance % 65535;
+    uint renderPriority = drawCallData.renderPriority;
+    uint invDistanceFromCameraUint = UINT_MAX - (uint)(distanceFromCamera / distanceAccuracy);
+    uint localInstanceID = drawCall.firstInstance % 65535;
     
     uint64_t sortKey = 0;
